v2/foundry-pipeline/requirements.txt
azure-identity>=1.15.0
requests>=2.31.0
azure-ai-projects>=2.0.0b4
